+++ packages/backend-plugin-api/src/services/definitions/RootLifecycleService.ts
@@ -23,4 +23,6 @@ import { LifecycleService } from './LifecycleService';
  *
  * @public
  */
-export interface RootLifecycleService extends LifecycleService {}
+export interface RootLifecycleService extends LifecycleService {
+  addBeforeShutdownHook(hook: () => void | Promise<void>): void;
+}