+++ plugins/scaffolder-backend-module-bitbucket-cloud/src/autocomplete/autocomplete.ts
@@ -25,7 +25,7 @@ export async function handleAutocompleteRequest({
   resource: string;
   token: string;
   context: Record<string, string>;
-}): Promise<{ results: { title: string }[] }> {
+}): Promise<{ results: { title?: string; id: string }[] }> {
   const client = BitbucketCloudClient.fromConfig({
     host: 'bitbucket.org',
     apiBaseUrl: 'https://api.bitbucket.org/2.0',
@@ -34,10 +34,12 @@ export async function handleAutocompleteRequest({
 
   switch (resource) {
     case 'workspaces': {
-      const results: { title: string }[] = [];
+      const results: { title?: string; id: string }[] = [];
 
       for await (const page of client.listWorkspaces().iteratePages()) {
-        const slugs = [...page.values!].map(p => ({ title: p.slug! }));
+        const slugs = [...page.values!].map(p => ({
+          id: p.slug!,
+        }));
         results.push(...slugs);
       }
 
@@ -47,12 +49,14 @@ export async function handleAutocompleteRequest({
       if (!context.workspace)
         throw new InputError('Missing workspace context parameter');
 
-      const results: { title: string }[] = [];
+      const results: { title?: string; id: string }[] = [];
 
       for await (const page of client
         .listProjectsByWorkspace(context.workspace)
         .iteratePages()) {
-        const keys = [...page.values!].map(p => ({ title: p.key! }));
+        const keys = [...page.values!].map(p => ({
+          id: p.key!,
+        }));
         results.push(...keys);
       }
 
@@ -64,14 +68,16 @@ export async function handleAutocompleteRequest({
           'Missing workspace and/or project context parameter',
         );
 
-      const results: { title: string }[] = [];
+      const results: { title?: string; id: string }[] = [];
 
       for await (const page of client
         .listRepositoriesByWorkspace(context.workspace, {
           q: `project.key="${context.project}"`,
         })
         .iteratePages()) {
-        const slugs = [...page.values!].map(p => ({ title: p.slug! }));
+        const slugs = [...page.values!].map(p => ({
+          id: p.slug!,
+        }));
         results.push(...slugs);
       }
 
@@ -83,12 +89,14 @@ export async function handleAutocompleteRequest({
           'Missing workspace and/or repository context parameter',
         );
 
-      const results: { title: string }[] = [];
+      const results: { title?: string; id: string }[] = [];
 
       for await (const page of client
         .listBranchesByRepository(context.repository, context.workspace)
         .iteratePages()) {
-        const names = [...page.values!].map(p => ({ title: p.name! }));
+        const names = [...page.values!].map(p => ({
+          id: p.name!,
+        }));
         results.push(...names);
       }