+++ plugins/search-backend-module-catalog/src/collators/defaultCatalogCollatorEntityTransformer.test.ts
@@ -147,5 +147,34 @@ describe('DefaultCatalogCollatorEntityTransformer', () => {
         owner: '',
       });
     });
+
+    it('sets text correctly when description is not provided', async () => {
+      const userEntityWithoutDescription = {
+        ...userEntity,
+        metadata: {
+          ...userEntity.metadata,
+          description: undefined,
+        },
+        spec: {
+          profile: {
+            ...userEntity.spec.profile,
+            email: undefined,
+          },
+        },
+      };
+
+      const document = defaultCatalogCollatorEntityTransformer(
+        userEntityWithoutDescription,
+      );
+
+      expect(document).toMatchObject({
+        title: userEntity.metadata.name,
+        text: userEntity.spec.profile.displayName,
+        namespace: 'default',
+        componentType: 'other',
+        lifecycle: '',
+        owner: '',
+      });
+    });
   });
 });