+++ packages/frontend-defaults/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @backstage/frontend-defaults
 
+## 0.1.4-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/errors@1.2.6-next.0
+  - @backstage/config@1.3.1-next.0
+  - @backstage/frontend-app-api@0.10.3-next.2
+  - @backstage/frontend-plugin-api@0.9.3-next.2
+  - @backstage/plugin-app@0.1.4-next.2
+
+## 0.1.4-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.3.0
+  - @backstage/errors@1.2.5
+  - @backstage/frontend-app-api@0.10.3-next.1
+  - @backstage/frontend-plugin-api@0.9.3-next.1
+  - @backstage/plugin-app@0.1.4-next.1
+
+## 0.1.4-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-app@0.1.4-next.0
+  - @backstage/frontend-plugin-api@0.9.3-next.0
+  - @backstage/config@1.3.0
+  - @backstage/errors@1.2.5
+  - @backstage/frontend-app-api@0.10.3-next.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 44b82da: The default config loader no longer requires `process.env.APP_CONFIG` to be set, allowing config to be read from other sources instead.
+- Updated dependencies
+  - @backstage/config@1.3.0
+  - @backstage/errors@1.2.5
+  - @backstage/frontend-app-api@0.10.1
+  - @backstage/frontend-plugin-api@0.9.1
+  - @backstage/plugin-app@0.1.2
+
 ## 0.1.2-next.2
 
 ### Patch Changes