+++ packages/techdocs-cli-embedded-app/CHANGELOG.md
@@ -1,5 +1,81 @@
 # techdocs-cli-embedded-app
 
+## 0.2.103-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.26.0-next.2
+  - @backstage/cli@0.29.3-next.2
+  - @backstage/app-defaults@1.5.15-next.2
+  - @backstage/catalog-model@1.7.2-next.0
+  - @backstage/config@1.3.1-next.0
+  - @backstage/core-app-api@1.15.3-next.1
+  - @backstage/core-components@0.16.2-next.2
+  - @backstage/core-plugin-api@1.10.2-next.0
+  - @backstage/integration-react@1.2.2-next.1
+  - @backstage/test-utils@1.7.3-next.1
+  - @backstage/theme@0.6.3-next.0
+  - @backstage/plugin-techdocs@1.11.3-next.2
+  - @backstage/plugin-techdocs-react@1.2.12-next.2
+
+## 0.2.103-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/cli@0.29.3-next.1
+  - @backstage/core-components@0.16.2-next.1
+  - @backstage/plugin-catalog@1.25.2-next.1
+  - @backstage/plugin-techdocs@1.11.3-next.1
+  - @backstage/app-defaults@1.5.15-next.1
+  - @backstage/catalog-model@1.7.1
+  - @backstage/config@1.3.0
+  - @backstage/core-app-api@1.15.3-next.0
+  - @backstage/core-plugin-api@1.10.1
+  - @backstage/integration-react@1.2.2-next.0
+  - @backstage/test-utils@1.7.3-next.0
+  - @backstage/theme@0.6.3-next.0
+  - @backstage/plugin-techdocs-react@1.2.12-next.1
+
+## 0.2.103-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/cli@0.29.3-next.0
+  - @backstage/core-app-api@1.15.3-next.0
+  - @backstage/plugin-techdocs@1.11.3-next.0
+  - @backstage/theme@0.6.3-next.0
+  - @backstage/app-defaults@1.5.15-next.0
+  - @backstage/catalog-model@1.7.1
+  - @backstage/config@1.3.0
+  - @backstage/core-components@0.16.2-next.0
+  - @backstage/core-plugin-api@1.10.1
+  - @backstage/integration-react@1.2.2-next.0
+  - @backstage/test-utils@1.7.3-next.0
+  - @backstage/plugin-catalog@1.25.2-next.0
+  - @backstage/plugin-techdocs-react@1.2.12-next.0
+
+## 0.2.102
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/cli@0.29.0
+  - @backstage/plugin-catalog@1.25.0
+  - @backstage/config@1.3.0
+  - @backstage/theme@0.6.1
+  - @backstage/plugin-techdocs@1.11.1
+  - @backstage/core-components@0.16.0
+  - @backstage/plugin-techdocs-react@1.2.10
+  - @backstage/core-app-api@1.15.2
+  - @backstage/app-defaults@1.5.13
+  - @backstage/catalog-model@1.7.1
+  - @backstage/core-plugin-api@1.10.1
+  - @backstage/integration-react@1.2.1
+  - @backstage/test-utils@1.7.1
+
 ## 0.2.102-next.3
 
 ### Patch Changes