+++ plugins/example-todo-list-common/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @internal/plugin-todo-list-common
 
+## 1.0.23-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.8.3-next.0
+
+## 1.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.8.2
+
 ## 1.0.21
 
 ### Patch Changes