+++ plugins/permission-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-permission-react
 
+## 0.4.29-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.3.1-next.0
+  - @backstage/core-plugin-api@1.10.2-next.0
+  - @backstage/plugin-permission-common@0.8.3-next.0
+
+## 0.4.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.3.0
+  - @backstage/plugin-permission-common@0.8.2
+  - @backstage/core-plugin-api@1.10.1
+
 ## 0.4.27
 
 ### Patch Changes